Fix identifier_exists false
With #55 we introduced a bug and BC break. If `identifierExists` is not set on the item (so is `null`) the feed should not contain `<g:identifier_exists>false</g:identifier_exists>`. That tag should be present only if `identifierExists` is exactly false.
But `if not item.identifierExists` will evaluate to true if `identifierExists` is null.